requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pandas>=1.5.0
requests>=2.28.0
opencv-python>=4.8.0
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=9.0.0
numpy>=1.24.0
python-dotenv>=1.0.0 